Add login and register pages, update dependencies, and refactor App component
