+++ frontend/src/Pages/App.jsx
@@ -1,7 +1,7 @@
 export default function Home() {
     return(
         <>
-        <p>hello</p>
+        
         </>
     )
 }
