+++ frontend/src/Pages/Login.jsx
@@ -1,0 +1,7 @@
+export default function Login() {
+    return (
+    <>
+    
+    </>
+    )
+}
