+++ frontend/src/Pages/Register.jsx
@@ -1,0 +1,6 @@
+export default function Register() {
+    return (
+        <>
+        </>
+    )
+}
