+++ package.json
@@ -1,0 +1,5 @@
+{
+	"dependencies": {
+		"react-error-boundary": "^5.0.0"
+	}
+}
